+++ runtime/src/main/java/com/instana/android/core/InstanaConfig.kt
@@ -119,6 +119,9 @@ class InstanaConfig
     var deleteOldBeacons:Boolean = false,
 ) {
 
+    //This need to be provided in Performance config If the n/w analytic is need to be
+    internal var enableNetworkUsageAnalytics: Boolean = false
+
     val breadcrumbsBufferSize: Int = 20
 
     /**